+++ client/web/src/enterprise/insights/pages/insights/creation/CreationRoutes.tsx
@@ -70,7 +70,6 @@ export const CreationRoutes: React.FunctionComponent<React.PropsWithChildren<Cre
 
             <Route
                 path={`${match.url}/group-results`}
-                exact={true}
                 render={() => (
                     <InsightCreationLazyPage
                         mode={InsightCreationPageType.Compute}