+++ client/web/src/org/openBeta/GettingStarted.tsx
@@ -7,7 +7,7 @@ import { RouteComponentProps } from 'react-router'
 
 import { ErrorAlert } from '@sourcegraph/branded/src/components/alerts'
 import { AuthenticatedUser } from '@sourcegraph/shared/src/auth'
-import { Link, LoadingSpinner, PageHeader, Badge, H3, Tooltip, Icon } from '@sourcegraph/wildcard'
+import { Link, LoadingSpinner, PageHeader, Badge, H3, Icon } from '@sourcegraph/wildcard'
 
 import { MarketingBlock } from '../../components/MarketingBlock'
 import { PageTitle } from '../../components/PageTitle'
@@ -372,23 +372,25 @@ export const OpenBetaGetStartedPage: React.FunctionComponent<React.PropsWithChil
                             <div className="d-flex  flex-0 justify-content-center align-items-center mb-3 flex-wrap">
                                 <div className={styles.membersList}>
                                     <div className={styles.avatarContainer}>
-                                        <Tooltip content={authenticatedUser.displayName || authenticatedUser.username}>
-                                            <UserAvatar size={36} className={styles.avatar} user={authenticatedUser} />
-                                        </Tooltip>
+                                        <UserAvatar
+                                            size={36}
+                                            className={styles.avatar}
+                                            user={authenticatedUser}
+                                            data-tooltip={authenticatedUser.displayName || authenticatedUser.username}
+                                        />
                                     </div>
                                     {otherMembers.length > 0 && (
                                         <div className={styles.avatarContainer}>
                                             <div className={classNames(styles.avatarEllipse)} />
                                             <div className={classNames(styles.avatarContainer, styles.secondAvatar)}>
-                                                <Tooltip
-                                                    content={otherMembers[0].displayName || otherMembers[0].username}
-                                                >
-                                                    <UserAvatar
-                                                        size={36}
-                                                        className={styles.avatar}
-                                                        user={otherMembers[0]}
-                                                    />
-                                                </Tooltip>
+                                                <UserAvatar
+                                                    size={36}
+                                                    className={styles.avatar}
+                                                    user={otherMembers[0]}
+                                                    data-tooltip={
+                                                        otherMembers[0].displayName || otherMembers[0].username
+                                                    }
+                                                />
                                             </div>
                                         </div>
                                     )}