+++ client/web/src/search/results/StreamingSearchResults.test.tsx
@@ -124,7 +124,7 @@ describe('StreamingSearchResults', () => {
         renderWrapper(<StreamingSearchResults {...defaultProps} streamSearch={() => of(COLLAPSABLE_SEARCH_RESULT)} />)
 
         expect(screen.getByTestId('search-result-expand-btn')).toHaveAttribute(
-            'data-test-tooltip-content',
+            'data-tooltip',
             'Show more matches on all results'
         )
 
@@ -135,7 +135,7 @@ describe('StreamingSearchResults', () => {
         userEvent.click(screen.getByTestId('search-result-expand-btn'))
 
         expect(screen.getByTestId('search-result-expand-btn')).toHaveAttribute(
-            'data-test-tooltip-content',
+            'data-tooltip',
             'Hide more matches on all results'
         )