+++ client/web/src/tree/FileDecorator.tsx
@@ -5,7 +5,6 @@ import classNames from 'classnames'
 import { FileDecoration } from 'sourcegraph'
 
 import { fileDecorationColorForTheme } from '@sourcegraph/shared/src/api/extension/api/decorations'
-import { Tooltip } from '@sourcegraph/wildcard'
 
 import styles from './FileDecorator.module.scss'
 
@@ -55,44 +54,44 @@ export const FileDecorator: React.FunctionComponent<React.PropsWithChildren<File
                             key={fileDecoration.uri + String(index)}
                         >
                             {fileDecoration.after && (
-                                <Tooltip content={fileDecoration.after.hoverMessage} placement="bottom">
-                                    <small
-                                        // eslint-disable-next-line react/forbid-dom-props
-                                        style={{
-                                            color: fileDecorationColorForTheme(
-                                                fileDecoration.after,
-                                                isLightTheme,
-                                                isActive
-                                            ),
-                                        }}
-                                        className={classNames(
-                                            'text-monospace font-weight-normal test-file-decoration-text',
-                                            styles.after,
-                                            isActive && styles.afterActive
-                                        )}
-                                    >
-                                        {fileDecoration.after.contentText}
-                                    </small>
-                                </Tooltip>
+                                <small
+                                    // eslint-disable-next-line react/forbid-dom-props
+                                    style={{
+                                        color: fileDecorationColorForTheme(
+                                            fileDecoration.after,
+                                            isLightTheme,
+                                            isActive
+                                        ),
+                                    }}
+                                    data-tooltip={fileDecoration.after.hoverMessage}
+                                    data-placement="bottom"
+                                    className={classNames(
+                                        'text-monospace font-weight-normal test-file-decoration-text',
+                                        styles.after,
+                                        isActive && styles.afterActive
+                                    )}
+                                >
+                                    {fileDecoration.after.contentText}
+                                </small>
                             )}
                             {fileDecoration.meter && (
-                                <Tooltip content={fileDecoration.meter.hoverMessage} placement="bottom">
-                                    <div>
-                                        <VisuallyHidden>{fileDecoration.meter.hoverMessage}</VisuallyHidden>
-                                        <meter
-                                            className={classNames('test-file-decoration-meter', styles.meter, {
-                                                'ml-2': !!fileDecoration.after,
-                                            })}
-                                            min={fileDecoration.meter.min}
-                                            low={fileDecoration.meter.low}
-                                            high={fileDecoration.meter.high}
-                                            max={fileDecoration.meter.max}
-                                            optimum={fileDecoration.meter.optimum}
-                                            value={fileDecoration.meter.value}
-                                            aria-hidden={true}
-                                        />
-                                    </div>
-                                </Tooltip>
+                                <div>
+                                    <VisuallyHidden>{fileDecoration.meter.hoverMessage}</VisuallyHidden>
+                                    <meter
+                                        className={classNames('test-file-decoration-meter', styles.meter, {
+                                            'ml-2': !!fileDecoration.after,
+                                        })}
+                                        min={fileDecoration.meter.min}
+                                        low={fileDecoration.meter.low}
+                                        high={fileDecoration.meter.high}
+                                        max={fileDecoration.meter.max}
+                                        optimum={fileDecoration.meter.optimum}
+                                        value={fileDecoration.meter.value}
+                                        data-tooltip={fileDecoration.meter.hoverMessage}
+                                        aria-hidden={true}
+                                        data-placement="bottom"
+                                    />
+                                </div>
                             )}
                         </div>
                     )