+++ enterprise/cmd/frontend/main.go
@@ -24,6 +24,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codemonitors"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/compute"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/dependencies"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/dotcom"
 	executor "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executorqueue"
 	licensing "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/licensing/init"
@@ -60,6 +61,7 @@ var initFunctions = map[string]EnterpriseInitializer{
 	"enterprise":     orgrepos.Init,
 	"notebooks":      notebooks.Init,
 	"compute":        compute.Init,
+	"dependencies":   dependencies.Init,
 }
 
 var codeIntelConfig = &codeintel.Config{}